Add new widget with list of all indexes to be displayed on Schema page. Fixes #19.
