+++ orienteer-core/src/main/java/org/orienteer/core/component/widget/AbstractIndexesWidget.html
@@ -2,4 +2,5 @@
     <form wicket:id="form">
         <table wicket:id="indexies" class="table table-condensed table-hover"></table>
     </form>
+    <wicket:child/>
 </wicket:extend>
