+++ orienteer-core/src/main/resources/org/orienteer/core/OrienteerWebApplication.properties
@@ -58,6 +58,8 @@ default.title=Default Title
 login.title=Login Page
 
 schema=Schema
+schema.all.indexes=All indexes
+
 
 class.list.title=List Classes
 class.name=Name
